requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cron-descriptor==1.4.5
Django==5.1.8
django-celery-beat>=2.5.0
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
kombu==5.5.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg-binary==3.2.6
PyJWT==2.9.0
pytest==8.3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
